# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorInfo:
    description: Detailed information about an error that occurred during an API request.
    properties:
      code:
        description: |-
          Code is a machine-readable error code identifying the specific error type.
          required: true
          example: RESOURCE_NOT_FOUND
        example: RESOURCE_NOT_FOUND
        type: string
      details:
        description: |-
          Details provides optional additional information about the error, such as validation failures.
          example: {"field": "command", "error": "cannot be empty"}
      message:
        description: |-
          Message is a human-readable description of the error.
          required: true
          example: The requested container was not found.
        example: The requested container was not found.
        type: string
    type: object
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse:
    description: Standard structure for returning errors from the API.
    properties:
      error:
        allOf:
        - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorInfo'
        description: |-
          Error contains the detailed error information.
          required: true
      meta:
        allOf:
        - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.MetadataResponse'
        description: |-
          Meta contains metadata about the response.
          required: true
      success:
        description: |-
          Success indicates if the request was successful (always false for errors).
          required: true
          example: false
        example: false
        type: boolean
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.MetadataResponse:
    properties:
      pagination:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.PaginationResponse'
      request_id:
        example: req-12345
        type: string
      timestamp:
        example: "2023-10-27T10:30:00Z"
        type: string
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.PaginatedResponse:
    properties:
      data:
        description: Should hold the slice of items
      meta:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.MetadataResponse'
      success:
        example: true
        type: boolean
    type: object
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.PaginationResponse:
    properties:
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total_items:
        example: 42
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse:
    properties:
      data:
        description: Use omitempty if data might be nil/empty
      message:
        description: Optional success message
        type: string
      meta:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.MetadataResponse'
      success:
        example: true
        type: boolean
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.VolumeCreateRequest:
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.VolumeResponse:
    description: Contains comprehensive details of a Docker volume, including configuration
      and usage.
    properties:
      containers:
        description: |-
          Containers lists the IDs of containers currently using this volume. Requires UsageData from Docker API.
          example: ["f7d9e8c7b6a5", "a1b2c3d4e5f6"]
        items:
          type: string
        type: array
      created_at:
        description: |-
          CreatedAt is the timestamp when the volume was created by Docker. Note: Docker API might return this as 'CreatedAt' string, needs parsing.
          required: true
          example: "2023-10-27T08:00:00Z"
        example: "2023-10-26T12:00:00Z"
        type: string
      driver:
        description: |-
          Driver is the volume driver used (e.g., local).
          required: true
          example: "local"
        example: local
        type: string
      id:
        description: |-
          ID is the internal database ID (if managed by the application).
          example: 12
        type: integer
      in_use:
        description: |-
          InUse indicates whether the volume is currently used by any containers. Requires UsageData from Docker API.
          example: true
        type: boolean
      labels:
        additionalProperties:
          type: string
        description: |-
          Labels are the labels applied to the volume.
          example: {"environment": "production", "backup": "true"}
        type: object
      mountpoint:
        description: |-
          Mountpoint is the path on the host where the volume data is stored.
          required: true
          example: "/var/lib/docker/volumes/my-app-data/_data"
        example: /var/lib/docker/volumes/my-app-data/_data
        type: string
      name:
        description: |-
          Name is the user-defined name of the volume.
          required: true
          example: "my-app-data"
        example: my-app-data
        type: string
      notes:
        description: |-
          Notes are user-defined notes stored in the application database.
          example: "Persistent data for the main database."
        type: string
      options:
        additionalProperties:
          type: string
        description: Options are driver-specific options for the volume.
        type: object
      scope:
        description: |-
          Scope indicates the scope of the volume (e.g., local, global).
          required: true
          example: "local"
        example: local
        type: string
      size:
        description: |-
          Size is the calculated size of the volume in bytes. Requires UsageData from Docker API.
          example: 104857600
        type: integer
      size_human:
        description: |-
          SizeHuman is the calculated size of the volume in a human-readable format.
          example: "100MB"
        type: string
      status:
        additionalProperties:
          type: string
        description: Status provides low-level status information about the volume
          (driver-specific).
        type: object
      updated_at:
        description: |-
          UpdatedAt is the timestamp when the volume record was last updated in the application database.
          example: "2023-10-27T08:10:00Z"
        type: string
      user_id:
        description: |-
          UserID is the ID of the user who owns/created this volume record in the application database.
          example: 1
        type: integer
      volume_id:
        description: |-
          VolumeID is the unique identifier assigned by Docker (often the same as Name).
          required: true
          example: "my-app-data"
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Docker Server Manager Go MCP.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Docker Server Manager Go MCP API
  version: "1.0"
paths:
...
  /volumes:
    get:
      description: Retrieves a paginated list of Docker volumes, optionally filtered.
      parameters:
      - default: 1
        description: Page number
        example: 1
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        example: 20
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - default: created_at
        description: Field to sort by (e.g., name, created_at, driver)
        example: name
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order (asc, desc)
        enum:
        - asc
        - desc
        example: asc
        in: query
        name: sort_order
        type: string
      - description: Filter by driver name
        example: local
        in: query
        name: driver
        type: string
      - description: Filter by volume name (exact match)
        example: my-app-data
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved volumes
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.VolumeResponse'
                  type: array
              type: object
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List volumes
      tags:
      - Volumes
    post:
      consumes:
      - application/json
      description: Creates a new Docker volume.
      parameters:
      - description: Volume Configuration
        in: body
        name: volume
        required: true
        schema:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.VolumeCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created volume" // Changed to 200 OK as we return
            the created object
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.VolumeResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "409":
          description: Volume name already exists
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new volume
      tags:
      - Volumes
  /volumes/{id}:
    delete:
      description: Removes a Docker volume by its name or database ID.
      parameters:
      - description: Volume Name or Database ID
        example: my-app-data
        in: path
        name: id
        required: true
        type: string
      - default: false
        description: Force removal of the volume
        example: true
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "204":
          description: Volume removed successfully
        "400":
          description: Invalid volume ID/name
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "404":
          description: Volume not found
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "409":
          description: Conflict (e.g., volume is in use)
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove a volume
      tags:
      - Volumes
    get:
      description: Retrieves detailed information about a specific volume by its name
        or database ID.
      parameters:
      - description: Volume Name or Database ID
        example: my-app-data
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved volume details
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.VolumeResponse'
              type: object
        "400":
          description: Invalid volume ID/name
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "404":
          description: Volume not found
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get volume details
      tags:
      - Volumes
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"